NH0046

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo fechas de operacion.

Este es un dataset de ejemplo que contiene informacion sobre estaciones meteorologicas, incluyendo ubicacion, tipo, y fechas de operacion.

Usage

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

get_data(estacion, destfile)

Format

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Un data frame con estas 35 variables:

id

ID de la estacion, datos categoricos.

fecha

Fecha de la medicion, datos temporales.

temperatura_abrigo_150cm

Temperatura medida en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_maxima

Temperatura maxima en abrigo a 150 cm, datos numericos.

temperatura_abrigo_150cm_minima

Temperatura minima en abrigo a 150 cm, datos numericos.

temperatura_intemperie_5cm_minima

Temperatura minima medida en intemperie a 5 cm del suelo, valores NA.

temperatura_intemperie_50cm_minima

Temperatura minima medida en intemperie a 50 cm del suelo, valores NA.

temperatura_suelo_5cm_media

Temperatura media medida en el suelo a 5 cm de profundidad, valores NA.

temperatura_suelo_10cm_media

Temperatura media medida en el suelo a 10 cm de profundidad, valores NA.

temperatura_inte_5cm

Temperatura medida a 5 cm en intemperie, valores NA.

temperatura_intemperie_150cm_minima

Temperatura minima en intemperie a 150 cm, valores NA.

humedad_suelo

Humedad del suelo, valores NA.

precipitacion_pluviometrica

Precipitacion pluviometrica en mm, datos numericos.

granizo

Indica presencia de granizo, valores NA.

nieve

Indica presencia de nieve, valores NA.

heliofania_efectiva

Heliofania efectiva, valores NA.

heliofania_relativa

Heliofania relativa, valores NA.

tesion_vapor_media

Tensión de vapor media, valores NA.

humedad_media

Humedad relativa media, valores NA.

humedad_media_8_14_20

Humedad relativa media registrada a las 8, 14 y 20 horas, valores NA.

rocio_medio

Punto de rocio medio, valores NA.

duracion_follaje_mojado

Duracion del follaje mojado, valores NA.

velocidad_viento_200cm_media

Velocidad media del viento a 200 cm de altura, valores NA.

direccion_viento_200cm

Direccion del viento a 200 cm de altura, valores NA.

velocidad_viento_1000cm_media

Velocidad media del viento a 1000 cm de altura, valores NA.

direccion_viento_1000cm

Direccion del viento a 1000 cm de altura, valores NA.

velocidad_viento_maxima

Velocidad maxima del viento registrada, valores NA.

presion_media

Presion atmosferica media, valores NA.

radiacion_global

Radiacion solar global medida en la estacion, valores NA.

radiacion_neta

Radiacion neta medida en la estacion, valores NA.

evaporacion_tanque

Evaporacion medida en tanque, valores NA.

evapotranspiracion_potencial

Evapotranspiracion potencial, valores NA.

profundidad_napa

Profundidad de la napa freatica, valores NA.

horas_frio

Cantidad de horas de frio registradas, valores numéricos.

unidad_frio

Unidad de medida para las horas de frio, valores NA.

Un data frame con 258 filas y 11 variables:

id

ID de la estacion, datos categoricos.

nombre

Nombre de la estacion, datos categoricos.

tipo

Tipo de estacion, posee datos categoricos.

localidad

Ciudad de la estacion, datos categoricos.

provincia

Provincia de la estacion, datos categoricos.

lat

Latitud de la ubicacion de la estacion, datos numericos.

lon

Longitud de la ubicacion de la estacion, datos numericos.

altura

Altura de la ubicacion de la estacion, datos numericos.

ubicacion

Ubicacion exacta de la estacion, datos categoricos.

desde

Fecha en la que inicio la estacion, datos temporales.

hasta

Fecha en la que finalizo la estacion, datos temporales

Source

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Los datos fueron obtenidos de los materiales otorgados en la clase de Programacion II.

Examples

#Cargar el dataset
data(NH0046)

#Ver las primeras filas del dataset
head(NH0046)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0046 1973-01-01                     27                              33.6
#> 2 NH0046 1973-01-02                     27.2                            34.9
#> 3 NH0046 1973-01-03                     25.2                            29.6
#> 4 NH0046 1973-01-04                     26.4                            32.6
#> 5 NH0046 1973-01-05                     26.0                            31  
#> 6 NH0046 1973-01-06                     20.4                            27.5
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <dbl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>,
#> #   nieve <dbl>, heliofania_efectiva <dbl>, heliofania_relativa <dbl>, …

#Cargar el dataset
data(NH0098)

#Ver las primeras filas del dataset
head(NH0098)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0098 1956-04-01                     13.6                            21.6
#> 2 NH0098 1956-04-02                     16.8                            23.6
#> 3 NH0098 1956-04-03                     19                              25.9
#> 4 NH0098 1956-04-04                     17.6                            23.6
#> 5 NH0098 1956-04-05                     20                              27.2
#> 6 NH0098 1956-04-06                     19.2                            24.7
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <dbl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>,
#> #   nieve <dbl>, heliofania_efectiva <dbl>, heliofania_relativa <dbl>, …

#Cargar el dataset
data(NH0437)

#Ver las primeras filas del dataset
head(NH0437)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0437 1970-08-01                     23.3                            31.2
#> 2 NH0437 1970-08-02                     26.6                            32.1
#> 3 NH0437 1970-08-03                     17.4                            22.9
#> 4 NH0437 1970-08-04                     11.0                            11.9
#> 5 NH0437 1970-08-05                     11.1                            12.6
#> 6 NH0437 1970-08-06                     13.6                            15.1
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>,
#> #   nieve <dbl>, heliofania_efectiva <dbl>, heliofania_relativa <dbl>, …

#Cargar el dataset
data(NH0472)

#Ver las primeras filas del dataset
head(NH0472)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0472 1961-07-01                     11.1                            13.4
#> 2 NH0472 1961-07-02                      5.5                            10.5
#> 3 NH0472 1961-07-03                      6                              10.8
#> 4 NH0472 1961-07-04                     NA                              NA  
#> 5 NH0472 1961-07-05                     12.4                            14.8
#> 6 NH0472 1961-07-06                     11                              16.5
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <dbl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>,
#> #   nieve <dbl>, heliofania_efectiva <dbl>, heliofania_relativa <dbl>, …

#Cargar el dataset
data(NH0910)

#Ver las primeras filas del dataset
head(NH0910)
#> # A tibble: 6 × 35
#>   id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>   <chr>  <date>                        <dbl>                           <dbl>
#> 1 NH0910 1975-01-01                     21                              31  
#> 2 NH0910 1975-01-02                     21.2                            28.5
#> 3 NH0910 1975-01-03                     24.2                            32.5
#> 4 NH0910 1975-01-04                     23.5                            31  
#> 5 NH0910 1975-01-05                     19                              22  
#> 6 NH0910 1975-01-06                     22.2                            27.5
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <lgl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <lgl>,
#> #   nieve <lgl>, heliofania_efectiva <lgl>, heliofania_relativa <lgl>, …

#Cargar el dataset
data(metadatos)

#Ver las primeras filas del dataset
head(metadatos)
#>        id                       nombre       tipo          localidad
#> 1 A872999    Instituto de Clima y Agua Nimbus THP         Hurlingham
#> 2 A872834         Lima - EEA San Pedro Nimbus THP               Lima
#> 3 A872823           Anguil -EEA Anguil Nimbus THP             Anguil
#> 4 A872822  Huinca Renanco - EEA Anguil Nimbus THP     Huinca Renanco
#> 5 A872821 Gral San Martin - EEA Anguil Nimbus THP General San Martin
#> 6 A872819       Miramar - EEA Balcarce Nimbus THP            Miramar
#>      provincia    lat    lon altura                                ubicacion
#> 1 Buenos Aires -90.00   0.00   22.0            Jardin Botanico Ragonese INTA
#> 2 Buenos Aires -34.08 -59.21   28.0                   Ruta 9 - Km 100 - Lima
#> 3     La Pampa -36.54 -63.99  165.0     Ruta Nacional Nº 5 Km. 580. CP: 6326
#> 4      Córdoba -34.77 -64.37  189.0           Ruta Nac. Nº35 Km.535 CP: 6270
#> 5     La Pampa -38.00 -63.52  177.9         Ruta Nac. Nº35 Km. 110. CP: 8206
#> 6 Buenos Aires -38.15 -57.99   12.0 Ruta Nº 77, km.20. CP:7607 Gral Alvarado
#>                       desde                     hasta
#> 1 2000-01-04T00:00:00Z[UTC] 2024-05-23T00:00:00Z[UTC]
#> 2 2010-08-25T00:00:00Z[UTC] 2024-03-23T00:00:00Z[UTC]
#> 3 2010-09-13T00:00:00Z[UTC] 2024-06-18T00:00:00Z[UTC]
#> 4 2010-09-06T00:00:00Z[UTC] 2022-11-02T00:00:00Z[UTC]
#> 5 2010-09-10T00:00:00Z[UTC] 2022-06-15T00:00:00Z[UTC]
#> 6 2010-11-11T00:00:00Z[UTC] 2024-02-01T00:00:00Z[UTC]